backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# OpenAI and AI dependencies
openai==1.3.7
tiktoken==0.5.2
tenacity==8.2.3

# Vector search and data processing
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.1.4

# HTTP and async
httpx==0.25.2
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Development and testing (optional for production)
pytest==7.4.3
pytest-asyncio==0.21.1 